Working; svg library included

+++ app/src/main/java/io/thepreviousone/openfloodii/activities/BaseActivity.java
@@ -2,10 +2,11 @@ package io.thepreviousone.openfloodii.activities;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.util.DisplayMetrics;
 import android.view.Window;
 import android.view.WindowManager;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import io.github.inflationx.viewpump.ViewPumpContextWrapper;
 
@@ -28,6 +29,6 @@ public abstract class BaseActivity extends AppCompatActivity {
         requestWindowFeature(Window.FEATURE_NO_TITLE);
         getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,
                 WindowManager.LayoutParams.FLAG_FULLSCREEN);
-        screenWidth = metrics.widthPixels / metrics.densityDpi;
+        screenWidth = (float) metrics.widthPixels / metrics.densityDpi;
     }
 }
